rev 1.1
Cubotino_m.py: Fixed bug at animation-start (when not found 6 different colours at centres).
GUI: Fixed minor bug at Warp_slicing slider (float instead of Integer).

+++ src/Cubotino_m_servos_GUI.py
@@ -590,7 +590,7 @@ def warp_f_get(val):
 
 def warp_s_get(val):
     global warp_slicing
-    warp_slicing = int(s_warp_s.get())         # Warping slincing
+    warp_slicing = float(s_warp_s.get())       # Warping slincing
     disp.show_on_display('WARP SLICING', str(warp_slicing), fs1=26, y2=75, fs2=30)  # feedback is printed to the display
     take_image(widgets_freeze=False)           # camera reads one frame
 ######################################################################################################################
